extract model map dialog
