+++ aqt/browser.py
@@ -21,6 +21,7 @@ from aqt.webview import AnkiWebView
 from aqt.toolbar import Toolbar
 from anki.consts import *
 from anki.sound import playFromText, clearAudioQueue
+from aqt.dialog.change_model import ChangeModelDialog
 
 COLOUR_SUSPENDED = "#FFFFB2"
 COLOUR_MARKED = "#D9B2E9"
@@ -977,7 +978,7 @@ where id in %s""" % ids2str(sf))
     def onChangeModel(self):
         nids = self.oneModelNotes()
         if nids:
-            ChangeModel(self, nids)
+            ChangeModelDialog(self, nids)
 
     def cram(self):
         return showInfo("not yet implemented")
@@ -1513,176 +1514,6 @@ update cards set usn=?, mod=?, did=? where id in """ + scids,
         except:
             return
         self.form.tableView.selectRow(row)
-
-# Change model dialog
-######################################################################
-
-class ChangeModel(QDialog):
-
-    def __init__(self, browser, nids):
-        QDialog.__init__(self, browser)
-        self.browser = browser
-        self.nids = nids
-        self.oldModel = browser.card.note().model()
-        self.form = aqt.forms.changemodel.Ui_Dialog()
-        self.form.setupUi(self)
-        self.setWindowModality(Qt.WindowModal)
-        self.setup()
-        restoreGeom(self, "changeModel")
-        addHook("reset", self.onReset)
-        addHook("currentModelChanged", self.onReset)
-        self.exec_()
-
-    def setup(self):
-        # maps
-        self.flayout = QHBoxLayout()
-        self.flayout.setContentsMargins(0,0,0,0)
-        self.fwidg = None
-        self.form.fieldMap.setLayout(self.flayout)
-        self.tlayout = QHBoxLayout()
-        self.tlayout.setContentsMargins(0,0,0,0)
-        self.twidg = None
-        self.form.templateMap.setLayout(self.tlayout)
-        if self.style().objectName() == "gtk+":
-            # gtk+ requires margins in inner layout
-            self.form.verticalLayout_2.setContentsMargins(0, 11, 0, 0)
-            self.form.verticalLayout_3.setContentsMargins(0, 11, 0, 0)
-        # model chooser
-        import aqt.modelchooser
-        self.oldModel = self.browser.col.models.get(
-            self.browser.col.db.scalar(
-                "select mid from notes where id = ?", self.nids[0]))
-        self.form.oldModelLabel.setText(self.oldModel['name'])
-        self.modelChooser = aqt.modelchooser.ModelChooser(
-            self.browser.mw, self.form.modelChooserWidget, label=False)
-        self.modelChooser.models.setFocus()
-        self.form.buttonBox.helpRequested.connect(self.onHelp)
-        self.modelChanged(self.browser.mw.col.models.current())
-        self.pauseUpdate = False
-
-    def onReset(self):
-        self.modelChanged(self.browser.col.models.current())
-
-    def modelChanged(self, model):
-        self.targetModel = model
-        self.rebuildTemplateMap()
-        self.rebuildFieldMap()
-
-    def rebuildTemplateMap(self, key=None, attr=None):
-        if not key:
-            key = "t"
-            attr = "tmpls"
-        map = getattr(self, key + "widg")
-        lay = getattr(self, key + "layout")
-        src = self.oldModel[attr]
-        dst = self.targetModel[attr]
-        if map:
-            lay.removeWidget(map)
-            map.deleteLater()
-            setattr(self, key + "MapWidget", None)
-        map = QWidget()
-        l = QGridLayout()
-        combos = []
-        targets = [x['name'] for x in dst] + [_("Nothing")]
-        indices = {}
-        for i, x in enumerate(src):
-            l.addWidget(QLabel(_("Change %s to:") % x['name']), i, 0)
-            cb = QComboBox()
-            cb.addItems(targets)
-            idx = min(i, len(targets)-1)
-            cb.setCurrentIndex(idx)
-            indices[cb] = idx
-            cb.currentIndexChanged.connect(
-                lambda i, cb=cb, key=key: self.onComboChanged(i, cb, key))
-            combos.append(cb)
-            l.addWidget(cb, i, 1)
-        map.setLayout(l)
-        lay.addWidget(map)
-        setattr(self, key + "widg", map)
-        setattr(self, key + "layout", lay)
-        setattr(self, key + "combos", combos)
-        setattr(self, key + "indices", indices)
-
-    def rebuildFieldMap(self):
-        return self.rebuildTemplateMap(key="f", attr="flds")
-
-    def onComboChanged(self, i, cb, key):
-        indices = getattr(self, key + "indices")
-        if self.pauseUpdate:
-            indices[cb] = i
-            return
-        combos = getattr(self, key + "combos")
-        if i == cb.count() - 1:
-            # set to 'nothing'
-            return
-        # find another combo with same index
-        for c in combos:
-            if c == cb:
-                continue
-            if c.currentIndex() == i:
-                self.pauseUpdate = True
-                c.setCurrentIndex(indices[cb])
-                self.pauseUpdate = False
-                break
-        indices[cb] = i
-
-    def getTemplateMap(self, old=None, combos=None, new=None):
-        if not old:
-            old = self.oldModel['tmpls']
-            combos = self.tcombos
-            new = self.targetModel['tmpls']
-        map = {}
-        for i, f in enumerate(old):
-            idx = combos[i].currentIndex()
-            if idx == len(new):
-                # ignore
-                map[f['ord']] = None
-            else:
-                f2 = new[idx]
-                map[f['ord']] = f2['ord']
-        return map
-
-    def getFieldMap(self):
-        return self.getTemplateMap(
-            old=self.oldModel['flds'],
-            combos=self.fcombos,
-            new=self.targetModel['flds'])
-
-    def cleanup(self):
-        remHook("reset", self.onReset)
-        remHook("currentModelChanged", self.onReset)
-        self.modelChooser.cleanup()
-        saveGeom(self, "changeModel")
-
-    def reject(self):
-        self.cleanup()
-        return QDialog.reject(self)
-
-    def accept(self):
-        # check maps
-        fmap = self.getFieldMap()
-        cmap = self.getTemplateMap()
-        if any(True for c in list(cmap.values()) if c is None):
-            if not askUser(_("""\
-Any cards mapped to nothing will be deleted. \
-If a note has no remaining cards, it will be lost. \
-Are you sure you want to continue?""")):
-                return
-        QDialog.accept(self)
-        self.browser.mw.checkpoint(_("Change Note Type"))
-        b = self.browser
-        b.mw.progress.start()
-        b.model.beginReset()
-        mm = b.mw.col.models
-        mm.change(self.oldModel, self.nids, self.targetModel, fmap, cmap)
-        b.onSearch(reset=False)
-        b.model.endReset()
-        b.mw.progress.finish()
-        b.mw.reset()
-        self.cleanup()
-
-    def onHelp(self):
-        openHelp("browsermisc")
 
 # Toolbar
 ######################################################################
